🎨 Align compiler APIs with merged C++ conventions
Move the adapter declarations to the mqt header namespace and apply the
merged include ordering and LLVM-style documentation comments. Preserve
the compiler and submission behavior across the rebase.
